Move credit card form in to separate component
