+++ src/Apps/Order/Routes/NewPayment/index.tsx
@@ -1,22 +1,11 @@
+import { NewPayment_me } from "__generated__/NewPayment_me.graphql"
 import { NewPayment_order } from "__generated__/NewPayment_order.graphql"
-import { NewPaymentRouteCreateCreditCardMutation } from "__generated__/NewPaymentRouteCreateCreditCardMutation.graphql"
 import { NewPaymentRouteSetOrderPaymentMutation } from "__generated__/NewPaymentRouteSetOrderPaymentMutation.graphql"
-import {
-  Address,
-  AddressChangeHandler,
-  AddressErrors,
-  AddressForm,
-  AddressTouched,
-  emptyAddress,
-} from "Apps/Order/Components/AddressForm"
-
 import { HorizontalPadding } from "Apps/Components/HorizontalPadding"
 import { ArtworkSummaryItemFragmentContainer as ArtworkSummaryItem } from "Apps/Order/Components/ArtworkSummaryItem"
-import { CreditCardInput } from "Apps/Order/Components/CreditCardInput"
 import { OrderStepper } from "Apps/Order/Components/OrderStepper"
 import { TransactionDetailsSummaryItemFragmentContainer as TransactionDetailsSummaryItem } from "Apps/Order/Components/TransactionDetailsSummaryItem"
 import { TwoColumnLayout } from "Apps/Order/Components/TwoColumnLayout"
-import { validateAddress } from "Apps/Order/Utils/formValidators"
 import { trackPageViewWrapper } from "Apps/Order/Utils/trackPageViewWrapper"
 import { track } from "Artsy/Analytics"
 import { CountdownTimer } from "Components/v2/CountdownTimer"
@@ -27,17 +16,11 @@ import { injectStripe, ReactStripeElements } from "react-stripe-elements"
 import createLogger from "Utils/logger"
 import { Media } from "Utils/Responsive"
 
+import { Button, Col, Flex, Join, Row, Spacer } from "@artsy/palette"
 import {
-  Button,
-  Checkbox,
-  Col,
-  Collapse,
-  Flex,
-  Join,
-  Row,
-  Serif,
-  Spacer,
-} from "@artsy/palette"
+  PaymentPicker,
+  PaymentPickerFragmentContainer,
+} from "Apps/Order/Components/PaymentPicker"
 import { Dialog, injectDialog } from "Apps/Order/Dialogs"
 import {
   CommitMutation,
@@ -54,6 +37,7 @@ export const ContinueButton = props => (
 export interface NewPaymentProps
   extends ReactStripeElements.InjectedStripeProps {
   order: NewPayment_order
+  me: NewPayment_me
   relay?: RelayRefetchProp
   router: Router
   route: RouteConfig
@@ -63,12 +47,7 @@ export interface NewPaymentProps
 }
 
 interface NewPaymentState {
-  hideBillingAddress: boolean
-  address: Address
-  addressErrors: AddressErrors
-  addressTouched: AddressTouched
-  stripeError: stripe.Error
-  isCreatingStripeToken: boolean
+  isGettingCreditCardId: boolean
 }
 
 const logger = createLogger("Order/Routes/NewPayment/index.tsx")
@@ -78,79 +57,31 @@ export class NewPaymentRoute extends Component<
   NewPaymentProps,
   NewPaymentState
 > {
+  paymentPicker = React.createRef<PaymentPicker>()
   state = {
-    hideBillingAddress: true,
-    stripeError: null,
-    address: this.startingAddress(),
-    addressErrors: {},
-    addressTouched: {},
-    isCreatingStripeToken: false,
-  }
-
-  startingAddress(): Address {
-    return {
-      ...emptyAddress,
-      country: "US",
-    }
-  }
-
-  get touchedAddress() {
-    return {
-      name: true,
-      country: true,
-      postalCode: true,
-      addressLine1: true,
-      addressLine2: true,
-      city: true,
-      region: true,
-      phoneNumber: true,
-    }
-  }
-
-  createStripeToken = async () => {
-    try {
-      this.setState({ isCreatingStripeToken: true })
-      const stripeBillingAddress = this.getStripeBillingAddress()
-      return await this.props.stripe.createToken(stripeBillingAddress)
-    } finally {
-      this.setState({ isCreatingStripeToken: false })
-    }
+    isGettingCreditCardId: false,
   }
 
   onContinue = async () => {
-    if (this.needsAddress()) {
-      const { errors, hasErrors } = validateAddress(this.state.address)
-      if (hasErrors) {
-        this.setState({
-          addressErrors: errors,
-          addressTouched: this.touchedAddress,
-        })
-        return
-      }
-    }
-
     try {
-      const stripeResult = await this.createStripeToken()
+      this.setState({ isGettingCreditCardId: true })
+      const result = await this.paymentPicker.current.getCreditCardId()
+      this.setState({ isGettingCreditCardId: false })
 
-      if (stripeResult.error) {
-        this.setState({ stripeError: stripeResult.error })
+      if (result.type === "invalid_form") {
         return
       }
 
-      const creditCardOrError = (await this.createCreditCard({
-        input: { token: stripeResult.token.id, oneTimeUse: true },
-      })).createCreditCard.creditCardOrError
-
-      if (creditCardOrError.mutationError) {
+      if (result.type === "error") {
         this.props.dialog.showErrorDialog({
-          message: creditCardOrError.mutationError.detail,
+          message: result.error,
         })
         return
       }
 
       const orderOrError = (await this.fixFailedPayment({
         input: {
-          creditCardId: creditCardOrError.creditCard.id,
+          creditCardId: result.creditCardId,
           offerId: this.props.order.lastOffer.id,
         },
       })).ecommerceFixFailedPayment.orderOrError
@@ -167,45 +98,11 @@ export class NewPaymentRoute extends Component<
     }
   }
 
-  handleChangeHideBillingAddress(hideBillingAddress: boolean) {
-    if (!hideBillingAddress) {
-      this.setState({
-        address: {
-          ...emptyAddress,
-          country: "US",
-        },
-      })
-    }
-
-    this.setState({ hideBillingAddress })
-  }
-
-  onAddressChange: AddressChangeHandler = (address, key) => {
-    const { errors } = validateAddress(address)
-    this.setState({
-      address,
-      addressErrors: {
-        ...this.state.addressErrors,
-        [key]: errors[key],
-      },
-      addressTouched: {
-        ...this.state.addressTouched,
-        [key]: true,
-      },
-    })
-  }
-
   render() {
     const { order, isCommittingMutation } = this.props
-    const {
-      stripeError,
-      address,
-      addressErrors,
-      addressTouched,
-      isCreatingStripeToken,
-    } = this.state
+    const { isGettingCreditCardId } = this.state
 
-    const isLoading = isCommittingMutation || isCreatingStripeToken
+    const isLoading = isCommittingMutation || isGettingCreditCardId
 
     return (
       <>
@@ -237,36 +134,12 @@ export class NewPaymentRoute extends Component<
                   </>
                 )}
                 <Join separator={<Spacer mb={3} />}>
-                  <Flex flexDirection="column">
-                    <Serif mb={1} size="3t" color="black100" lineHeight="1.1em">
-                      Credit card
-                    </Serif>
-                    <CreditCardInput
-                      error={stripeError}
-                      onChange={response => {
-                        this.setState({ stripeError: response.error })
-                      }}
-                    />
-                  </Flex>
-
-                  {!this.isPickup() && (
-                    <Checkbox
-                      selected={this.state.hideBillingAddress}
-                      onSelect={this.handleChangeHideBillingAddress.bind(this)}
-                    >
-                      Billing and shipping addresses are the same
-                    </Checkbox>
-                  )}
-                  <Collapse open={this.needsAddress()}>
-                    <AddressForm
-                      value={address}
-                      errors={addressErrors}
-                      touched={addressTouched}
-                      onChange={this.onAddressChange}
-                      billing
-                    />
-                  </Collapse>
-
+                  <PaymentPickerFragmentContainer
+                    order={order}
+                    me={this.props.me}
+                    commitMutation={this.props.commitMutation}
+                    innerRef={this.paymentPicker}
+                  />
                   <Media greaterThan="xs">
                     <ContinueButton
                       onClick={this.onContinue}
@@ -298,60 +171,6 @@ export class NewPaymentRoute extends Component<
         </HorizontalPadding>
       </>
     )
-  }
-
-  private getStripeBillingAddress(): stripe.TokenOptions {
-    const selectedBillingAddress = (this.needsAddress()
-      ? this.state.address
-      : this.props.order.requestedFulfillment) as Address
-    const {
-      name,
-      addressLine1,
-      addressLine2,
-      city,
-      region,
-      postalCode,
-      country,
-    } = selectedBillingAddress
-    return {
-      name,
-      address_line1: addressLine1,
-      address_line2: addressLine2,
-      address_city: city,
-      address_state: region,
-      address_zip: postalCode,
-      address_country: country,
-    }
-  }
-
-  createCreditCard(
-    variables: NewPaymentRouteCreateCreditCardMutation["variables"]
-  ) {
-    return this.props.commitMutation<NewPaymentRouteCreateCreditCardMutation>({
-      variables,
-      mutation: graphql`
-        mutation NewPaymentRouteCreateCreditCardMutation(
-          $input: CreditCardInput!
-        ) {
-          createCreditCard(input: $input) {
-            creditCardOrError {
-              ... on CreditCardMutationSuccess {
-                creditCard {
-                  id
-                }
-              }
-              ... on CreditCardMutationFailure {
-                mutationError {
-                  type
-                  message
-                  detail
-                }
-              }
-            }
-          }
-        }
-      `,
-    })
   }
 
   fixFailedPayment(
@@ -422,14 +241,6 @@ export class NewPaymentRoute extends Component<
     }
   }
 
-  private isPickup = () => {
-    return this.props.order.requestedFulfillment.__typename === "Pickup"
-  }
-
-  private needsAddress = () => {
-    return this.isPickup() || !this.state.hideBillingAddress
-  }
-
   artistId() {
     return get(
       this.props.order,
@@ -451,34 +262,15 @@ export const NewPaymentFragmentContainer = createFragmentContainer(
     injectStripe(trackPageViewWrapper(injectDialog(NewPaymentRoute)))
   ),
   {
+    me: graphql`
+      fragment NewPayment_me on Me {
+        ...PaymentPicker_me
+      }
+    `,
     order: graphql`
       fragment NewPayment_order on Order {
         id
         mode
-        creditCard {
-          name
-          street1
-          street2
-          city
-          state
-          country
-          postal_code
-        }
-        requestedFulfillment {
-          __typename
-          ... on Ship {
-            name
-            addressLine1
-            addressLine2
-            city
-            region
-            country
-            postalCode
-          }
-          ... on Pickup {
-            fulfillmentType
-          }
-        }
         stateExpiresAt
         lineItems {
           edges {
@@ -499,6 +291,7 @@ export const NewPaymentFragmentContainer = createFragmentContainer(
             note
           }
         }
+        ...PaymentPicker_order
         ...ArtworkSummaryItem_order
         ...TransactionDetailsSummaryItem_order
       }
