+++ src/Apps/Order/Routes/__tests__/Payment.test.tsx
@@ -75,6 +75,7 @@ describe("Payment", () => {
     Component: PaymentFragmentContainer,
     defaultData: {
       order: testOrder,
+      me: { creditCards: { edges: [] } },
     },
     defaultMutationResults: {
       ...creatingCreditCardSuccess,
@@ -82,6 +83,9 @@ describe("Payment", () => {
     },
     query: graphql`
       query PaymentTestQuery {
+        me {
+          ...Payment_me
+        }
         order: ecommerceOrder(id: "unused") {
           ...Payment_order
         }
@@ -115,7 +119,9 @@ describe("Payment", () => {
   beforeEach(() => {
     mockPostEvent.mockReset()
     createTokenMock.mockReset()
-    createTokenMock.mockImplementation(() => Promise.resolve())
+    createTokenMock.mockImplementation(() =>
+      Promise.resolve({ error: "bad error" })
+    )
   })
 
   it("always shows the billing address form without checkbox when the user selected 'pick' shipping option", async () => {
